Move filter button component into buttons folders
